Refactored the method that manages the captcha handling to a service. Modified the captcha appearance flow using the userValidatedByDomain.dataValidation.allowAnonymousAction flag passed as  parameter in POST request.
