+++ grails-app/controllers/kuorum/politician/CampaignController.groovy
@@ -7,6 +7,7 @@ import grails.plugin.springsecurity.annotation.Secured
 import groovy.time.TimeCategory
 import groovyx.net.http.RESTClient
 import kuorum.KuorumFile
+import kuorum.captcha.CaptchaService
 import kuorum.core.FileType
 import kuorum.core.customDomain.CustomDomainResolver
 import kuorum.core.exception.KuorumException
@@ -73,8 +74,9 @@ class CampaignController {
     CampaignService campaignService
     SearchSolrService searchSolrService
 
-    @Value('${recaptcha.providers.google.secretKey}')
-    String RECAPTCHA_SECRET
+    CaptchaService captchaService
+
+
 
     def show() {
         String viewerUid = cookieUUIDService.buildUserUUID()
@@ -522,26 +524,17 @@ class CampaignController {
                 pendingValidations: kuorumUserService.getPendingValidations(userValidationRSDTO, userSession)
         ] as JSON)
     }
-    //TODO same method logic in registerController, do refactor
-    def verifyCaptcha() {
-        String secretKey = RECAPTCHA_SECRET
-        String responseCaptcha = params['g-recaptcha-response']
-        String path = "/recaptcha/api/siteverify"
-        def query = [secret: secretKey, response: responseCaptcha]
-        RESTClient mailKuorumServices = new RESTClient("https://www.google.com")
-        def response = mailKuorumServices.get(path: path,
-                headers: ["User-Agent": "Kuorum Web"],
-                query: query,
-                requestContentType: groovyx.net.http.ContentType.JSON)
 
-        log.info("Checking CAPTCHA :: Google response - ${response.data.hostname} || domain : ${CustomDomainResolver.domain}")
-        if (!response.data.success && response.data.hostname != CustomDomainResolver.domain) return false else return true
-    }
 
-    // ajax/validation/domain-valid-phone-sendSms
     def validateUserPhoneSendSMS(DomainUserPhoneValidationCommand domainUserPhoneValidationCommand) {
-        boolean isCaptchaSolved = verifyCaptcha();
-        if (domainUserPhoneValidationCommand.hasErrors() || !isCaptchaSolved) {
+        def isCaptchaVerified = false
+        def isCaptchaNecessary = Boolean.parseBoolean(params['allowAnonymousAction']);
+        if (isCaptchaNecessary) {
+            isCaptchaVerified = captchaService.verifyCaptcha(params['g-recaptcha-response']);
+        } else {
+            isCaptchaVerified = true;
+        }
+        if (domainUserPhoneValidationCommand.hasErrors() || !isCaptchaVerified) {
             render([validated: false, success: false, hash: null, validationPhoneNumberPrefix: null, validationPhoneNumber: null, msg: g.message(code: 'kuorum.web.commands.profile.DomainUserPhoneValidationCommand.phoneNumber.invalidPhone')] as JSON)
             return;
         }
